Trying to use local library and mopidy core listeners
This is a work in progress commit. I have trouble setting up a custom
local library for the audiobook extensions. Currently it is using the local
library and i have changed local to audiobook at some places, but the uris
are still passed around as local. Atleast thats how the core.listener
intercepts them. Instead of a listener I would rather want to use the playbackprovider.play function to reimplement a playmethod for audiobooks.